+++ services/nutrition/src/main/resources/db/migration/V0_0_1__INIT_PG.sql
@@ -11,15 +11,15 @@ create table IF NOT EXISTS nutrition.history
     login            VARCHAR(255) not null,
     date             date         not null,
     breakfast        bigint,
-    breakfast_weight bigint,
+    breakfast_weight int,
     lunch            bigint,
-    lunch_weight     bigint,
+    lunch_weight     int,
     dinner           bigint,
-    dinner_weight    bigint,
-    total_tdee       int,
-    total_protein    int,
-    total_fat        int,
-    total_carbs      int
+    dinner_weight    int,
+    total_tdee       double precision,
+    total_protein    double precision,
+    total_fat        double precision,
+    total_carbs      double precision
 );
 
 create index IF NOT EXISTS history_search on nutrition.history (login, date)
